+++ app/src/androidTest/java/com/sdp/movemeet/Navigation/ChatActivityNavigationTest.java
@@ -22,7 +22,7 @@ import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.contrib.DrawerMatchers.isClosed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 
-public class UploadActivityTest {
+public class ChatActivityNavigationTest {
     @Rule
     public ActivityScenarioRule<HomeScreenActivity> testRule = new ActivityScenarioRule<>(HomeScreenActivity.class);
 
@@ -43,7 +43,7 @@ public class UploadActivityTest {
             Thread.sleep(500);
         }catch(Exception e){}
 
-        onView(withId(R.id.nav_view)).perform(NavigationViewActions.navigateTo(R.id.nav_add_activity));
+        onView(withId(R.id.nav_view)).perform(NavigationViewActions.navigateTo(R.id.nav_chat));
 
         try{
             Thread.sleep(500);
@@ -57,37 +57,37 @@ public class UploadActivityTest {
     }
 
     @Test
-    public void UploadActivityToProfileActivity() {
+    public void ChatActivityToProfileActivity() {
         onView(withId(R.id.nav_view)).perform(NavigationViewActions.navigateTo(R.id.nav_edit_profile));
         logout();
     }
 
     @Test
-    public void UploadActivityToActivityUpload() {
+    public void ChatActivityToActivityUpload() {
         onView(withId(R.id.nav_view)).perform(NavigationViewActions.navigateTo(R.id.nav_add_activity));
         logout();
     }
 
     @Test
-    public void UploadActivityToStartActivity() {
+    public void ChatActivityToStartActivity() {
         onView(withId(R.id.nav_view)).perform(NavigationViewActions.navigateTo(R.id.nav_start_activity));
         logout();
     }
 
     @Test
-    public void UploadActivityGotoHome() {
+    public void ChatActivityGotoHome() {
         onView(withId(R.id.nav_view)).perform(NavigationViewActions.navigateTo(R.id.nav_home));
         logout();
     }
 
     @Test
-    public void UploadActivityGotoChat() {
+    public void ChatActivityGotoChat() {
         onView(withId(R.id.nav_view)).perform(NavigationViewActions.navigateTo(R.id.nav_chat));
         logout();
     }
 
     @Test
-    public void UploadActivity_logout() {
+    public void ChatActivity_logout() {
         onView(withId(R.id.nav_view)).perform(NavigationViewActions.navigateTo(R.id.nav_logout));
     }
